remove all warnings

+++ week2/listwork1.cpp
@@ -15,18 +15,18 @@ int main()
         {
             l.push_front(k);
         }
-        else if(t == 'D' && (k >= 1 && k <= 20))
+        else if (t == 'D' && (k >= 1 && k <= 20))
         {
-            if (l.size() < k)
+            if ((int)l.size() < k)
             {
                 continue;
             }
-            list<int>:: iterator it = l.begin();
-            advance(it,k-1);
+            list<int>::iterator it = l.begin();
+            advance(it, k - 1);
             l.erase(it);
         }
     }
-    for (auto const& i : l) 
+    for (auto const &i : l)
     {
         std::cout << i << "\n";
     }
